requirements.txt
streamlit
requests
beautifulsoup4
wordcloud
Pillow 